- Moved the Dust storage to the UTXOManager
- Keep the collected dust stats during white-flag internal and do not apply them on confirmation
